Add IUCV to communicate with VM to replace SSH
When deploy VM, put IUCV related file to VM.

+++ nova/virt/zvm/utils.py
@@ -624,8 +624,11 @@ def xdsh(node, commands):
     return res_dict
 
 
-def punch_file(node, fn, fclass):
-    body = [" ".join(['--punchfile', fn, fclass, get_host()])]
+def punch_file(node, fn, fclass, remote_host=None):
+    if remote_host:
+        body = [" ".join(['--punchfile', fn, fclass, remote_host])]
+    else:
+        body = [" ".join(['--punchfile', fn, fclass])]
     url = get_xcat_url().chvm('/' + node)
 
     try:
@@ -636,14 +639,15 @@ def punch_file(node, fn, fclass):
             LOG.error(_('Punch file to %(node)s failed: %(msg)s') %
                       {'node': node, 'msg': emsg})
     finally:
-        os.remove(fn)
+        if remote_host:
+            os.remove(fn)
 
 
 def punch_adminpass_file(instance_path, instance_name, admin_password,
                          linuxdist):
     adminpass_fn = ''.join([instance_path, '/adminpwd.sh'])
     _generate_adminpass_file(adminpass_fn, admin_password, linuxdist)
-    punch_file(instance_name, adminpass_fn, 'X')
+    punch_file(instance_name, adminpass_fn, 'X', get_host())
 
 
 def punch_xcat_auth_file(instance_path, instance_name):
@@ -651,7 +655,65 @@ def punch_xcat_auth_file(instance_path, instance_name):
     mn_pub_key = get_mn_pub_key()
     auth_fn = ''.join([instance_path, '/xcatauth.sh'])
     _generate_auth_file(auth_fn, mn_pub_key)
-    punch_file(instance_name, auth_fn, 'X')
+    punch_file(instance_name, auth_fn, 'X', get_host())
+
+
+def _generate_iucv_cmd_file(iucv_cmd_file_path, cmd):
+    lines = ['#!/bin/bash\n', cmd]
+    with open(iucv_cmd_file_path, 'w') as f:
+        f.writelines(lines)
+
+
+def punch_iucv_file(os_ver, zhcp, xcat_iucv_path, instance_name):
+    """put iucvserver and iucvserverd files to reader."""
+    # generate iucvserver file
+    iucv_server_fn = ''.join([xcat_iucv_path, '/iucvserver'])
+    # generate iucvserver service and script file
+    iucv_service_name = ''
+    iucv_service_path = ''
+    start_iucv_service_sh = ''
+    (distro, release) = dist.ListDistManager().parse_dist(os_ver)
+    if((distro == "rhel" and release < '7') or (
+                 distro == "sles" and release < '12')):
+        iucv_serverd_fn = ''.join([xcat_iucv_path, '/iucvserverd'])
+        iucv_service_name = 'iucvserverd'
+        iucv_service_path = '/etc/init.d/' + iucv_service_name
+        start_iucv_service_sh = 'chkconfig --add iucvserverd\n'
+        start_iucv_service_sh += 'service iucvserverd  start\n'
+    else:
+        iucv_serverd_fn = ''.join([xcat_iucv_path, '/iucvserverd.service'])
+        iucv_service_name = 'iucvserverd.service'
+        start_iucv_service_sh = 'systemctl enable iucvserver.service\n'
+        start_iucv_service_sh += 'systemctl start iucvserver.service\n'
+        if(distro == "rhel" and release >= '7'):
+            iucv_service_path = '/lib/systemd/system/' + iucv_service_name
+        if(distro == "sles" and release >= '12'):
+            iucv_service_path = '/usr/lib/systemd/system/' + iucv_service_name
+
+    iucv_cmd_file_path = '/tmp/iucv_cmd_file_path.sh'
+    cmd = '\n'.join((
+        "spoolid=`vmur li | awk '/iucvserver/{print \$2}'|tail -1",
+        "vmur re -f $spoolid /usr/bin/iucvserver",
+        "spoolid=`vmur li | awk '/%s/{print \$2}'|tail -1`" %
+                                                  iucv_service_name,
+        "vmur re -f $spoolid %s" % iucv_service_path,
+        "echo %s >/tmp/authorized_userid 2>&1" % zhcp,
+        start_iucv_service_sh
+        ))
+    _generate_iucv_cmd_file(iucv_cmd_file_path, cmd)
+
+    punch_file(instance_name, iucv_server_fn, 'X')
+    punch_file(instance_name, iucv_serverd_fn, 'X')
+    punch_file(instance_name, iucv_cmd_file_path, 'X', get_host())
+
+
+def punch_iucv_authorized_file(instance_name, zhcp):
+    cmd = '\n'.join((
+        "echo %s >/tmp/authorized_userid 2>&1" % zhcp
+        ))
+    iucv_cmd_file_path = '/tmp/iucv_cmd_file_path.sh'
+    _generate_iucv_cmd_file(iucv_cmd_file_path, cmd)
+    punch_file(instance_name, iucv_cmd_file_path, 'X', get_host())
 
 
 def process_eph_disk(instance_name, vdev=None, fmt=None, mntdir=None):
@@ -679,13 +741,13 @@ def aemod_handler(instance_name, func_name, parms):
 
 
 def punch_configdrive_file(transportfiles, instance_name):
-    punch_file(instance_name, transportfiles, 'X')
+    punch_file(instance_name, transportfiles, 'X', get_host())
 
 
 def punch_zipl_file(instance_path, instance_name, lun, wwpn, fcp, volume_meta):
     zipl_fn = ''.join([instance_path, '/ziplset.sh'])
     _generate_zipl_file(zipl_fn, lun, wwpn, fcp, volume_meta)
-    punch_file(instance_name, zipl_fn, 'X')
+    punch_file(instance_name, zipl_fn, 'X', get_host())
 
 
 def generate_vdev(base, offset=1):
